backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-chess==1.999
stockfish==3.28.0
openai==1.3.5
pydantic==2.5.0
pydantic-core==2.14.1
python-dotenv==1.0.0